+++ config/routes/frontend_routes.rb
@@ -38,7 +38,7 @@ namespace :frontend, path: '', constraints: ->(req) { req.subdomain.blank? || !%
 
   get 'roadmap' => 'base#index'
   get 'roadmap/changes' => 'base#index'
-  get 'roadmap/releases' => 'base#index'
+  get 'roadmap/ships' => 'base#index'
 
   get 'impressum' => 'base#index'
   get 'privacy-policy' => 'base#index'